+++ jest.config.js
@@ -34,10 +34,10 @@ module.exports = {
   },
   coverageThreshold: {
     "global": {
-      "branches": 80,
-      "functions": 95,
-      "lines": 95,
-      "statements": 95
+      "branches": 75,
+      "functions": 90,
+      "lines": 90,
+      "statements": 90
     }
   },
   testEnvironment: "jsdom",