+++ src/display/Spacer/props.ts
@@ -1,0 +1,10 @@
+/**
+ * This component using bootstrap lib https://reactstrap.github.io/components/toasts/
+ * @example
+ * <Spacer size="xs" />
+ */
+
+export interface SpacerProps {
+    size: "sm" | "md" | "lg" | "xlg"
+}
+